+++ XCDYouTubeKit/XCDYouTubeVideoOperation.m
@@ -187,7 +187,7 @@ typedef NS_ENUM(NSUInteger, XCDYouTubeRequestType) {
 	if (self.webpage.isAgeRestricted)
 	{
 		NSString *eurl = [@"https://youtube.googleapis.com/v/" stringByAppendingString:self.videoIdentifier];
-		NSString *sts = [self.embedWebpage.playerConfiguration[@"sts"] description] ?: @"";
+		NSString *sts = [self.embedWebpage.playerConfiguration[@"sts"] description] ?: [self.webpage.playerConfiguration[@"sts"] description] ?: @"";
 		NSDictionary *query = @{ @"video_id": self.videoIdentifier, @"hl": self.languageIdentifier, @"eurl": eurl, @"sts": sts};
 		NSString *queryString = XCDQueryStringWithDictionary(query, NSUTF8StringEncoding);
 		NSURL *videoInfoURL = [NSURL URLWithString:[@"https://www.youtube.com/get_video_info?" stringByAppendingString:queryString]];